LLMShield-Backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.0.1
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
motor==3.7.1
passlib==1.7.4
Pillow==10.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.14.1
pyotp==2.8.0
pypng==0.20220715.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
qrcode==7.4.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
GitPython==3.1.40

langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0


